refactor: extract capability contracts and postgres primitives (#360)
