+++ architecture/modules/graph-contract.json
@@ -1,0 +1,19 @@
+{
+  "schema_version": 1,
+  "capability": "graph-contract",
+  "completed_issues": [],
+  "go": {
+    "units": [
+      {
+        "id": "github.com/markhuangai/dense-mem/internal/graph/contract",
+        "role": "port",
+        "visibility": "public"
+      }
+    ]
+  },
+  "browser": {
+    "units": []
+  },
+  "exceptions": [],
+  "workers": []
+}
