+++ architecture/modules/postgres-lock-admission.json
@@ -1,0 +1,19 @@
+{
+  "schema_version": 1,
+  "capability": "postgres-lock-admission",
+  "completed_issues": [],
+  "go": {
+    "units": [
+      {
+        "id": "github.com/markhuangai/dense-mem/internal/storage/postgres/lockadmission",
+        "role": "postgres_infrastructure",
+        "visibility": "private"
+      }
+    ]
+  },
+  "browser": {
+    "units": []
+  },
+  "exceptions": [],
+  "workers": []
+}
