+++ architecture/ownership.v1.json
@@ -96,6 +96,7 @@
     "architecture/modules/evidence-conflict-application.json",
     "architecture/modules/full-text-policy.json",
     "architecture/modules/graph-application.json",
+    "architecture/modules/graph-contract.json",
     "architecture/modules/http-error-policy.json",
     "architecture/modules/http-transport.json",
     "architecture/modules/mcp-transport.json",
@@ -108,21 +109,27 @@
     "architecture/modules/offline-evaluation.json",
     "architecture/modules/ownership-policy.json",
     "architecture/modules/postgres-adapter.json",
+    "architecture/modules/postgres-graph-read.json",
+    "architecture/modules/postgres-lock-admission.json",
     "architecture/modules/postgres-storage-adapter.json",
     "architecture/modules/process-coordination-adapter.json",
     "architecture/modules/prompt-policy.json",
     "architecture/modules/recall-quality-policy.json",
+    "architecture/modules/recall-contract.json",
     "architecture/modules/redis-coordination-adapter.json",
     "architecture/modules/remember-application.json",
     "architecture/modules/remember-assessor-port.json",
     "architecture/modules/remember-assessor-provider.json",
+    "architecture/modules/search-contract.json",
     "architecture/modules/request-context-policy.json",
     "architecture/modules/semantic-write-application.json",
+    "architecture/modules/semanticwrite-contract.json",
     "architecture/modules/server-composition.json",
     "architecture/modules/shared-domain.json",
     "architecture/modules/skill-pack-application.json",
     "architecture/modules/sse-transport.json",
     "architecture/modules/strict-json-policy.json",
+    "architecture/modules/trace-contract.json",
     "architecture/modules/tool-application-api.json",
     "architecture/modules/tool-registry-application-api.json",
     "architecture/modules/user-portal.json"
