+++ internal/graph/contract/query_test.go
@@ -1,0 +1,12 @@
+package contract
+
+import (
+	"reflect"
+	"testing"
+)
+
+func TestQueryDoesNotExposeDerivedSpaceScope(t *testing.T) {
+	if _, ok := reflect.TypeOf(Query{}).FieldByName("spaceID"); ok {
+		t.Fatal("graph query must not expose adapter-derived space scope")
+	}
+}
