+++ internal/service/memoryservice/recall_contract_test.go
@@ -1,0 +1,5 @@
+package memoryservice
+
+import recallcontract "github.com/markhuangai/dense-mem/internal/recall/contract"
+
+var _ recallcontract.Service = (*recallService)(nil)
