+++ internal/trace/contract/input_test.go
@@ -1,0 +1,12 @@
+package contract
+
+import (
+	"reflect"
+	"testing"
+)
+
+func TestInputDoesNotExposeDerivedSpaceScope(t *testing.T) {
+	if _, ok := reflect.TypeOf(Input{}).FieldByName("spaceID"); ok {
+		t.Fatal("trace input must not expose adapter-derived space scope")
+	}
+}
